+++ lib/signale.dart
@@ -1,8 +1,10 @@
 library signale;
 
 import 'dart:async';
+import 'dart:math';
 
 import 'package:signale/src/foundation/log_level.dart';
+import 'package:signale/src/foundation/random.dart';
 import 'package:signale/src/interface/print_able.dart';
 import 'package:signale/src/simple/default_print.dart';
 export 'src/interface/print_able.dart';